Added edit profile and added a basic wall page without functionality.

+++ database/migrations/2020_04_14_085844_create_followers_table.php
@@ -14,8 +14,8 @@ class CreateFollowersTable extends Migration
     public function up()
     {
         Schema::create('followers', function (Blueprint $table) {
-            $table->integer('follower_id')->unsigned();
-            $table->integer('leader_id')->unsigned();
+            $table->unsignedBigInteger('follower_id')->unsigned();
+            $table->unsignedBigInteger('leader_id')->unsigned();
             $table->timestamps();
             $table->foreign('follower_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('leader_id')->references('id')->on('users')->onDelete('cascade');
